aqi-karachi/requirements.txt
# Core Requirements for Python 3.12
streamlit==1.30.0
pymongo==4.6.1
pandas==2.1.4
numpy==1.26.0
plotly==5.18.0

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2

# Time Series Analysis
statsmodels==0.14.1

# Data Processing & Visualization
seaborn==0.13.0
matplotlib==3.8.2

# API & Web (updated versions)
requests==2.31.0
python-dotenv==1.0.0
openmeteo-requests==1.1.0
requests-cache==1.1.0
retry-requests==2.0.0  # Updated version

# Scheduling & Automation
schedule==1.2.0

# Development & Testing
jupyter==1.0.0
ipython==8.18.1
notebook==7.0.6
pytest==7.4.3